resolve: factor out resolve imports to its own module
